+++ shuup/admin/static_src/base/js/quick_add_related_objects.js
@@ -7,67 +7,114 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+"use strict";
+
 window.addToSelect2 = function addToSelect2(target, value, name) {
-    const newOption = new Option(name, value, true, true);
+    var newOption = new Option(name, value, true, true);
     $("select[name='" + target + "']").append(newOption).trigger("change");
 
     // Product update or create
     if (target.includes("primary_category")) {
-        const categories = $("select[name*='categories']");
+        var categories = $("select[name*='categories']");
         if (categories.length > 0) {
-            const newOption = new Option(name, value, true, true);
+            newOption = new Option(name, value, true, true);
             categories.append(newOption).trigger("change");
         }
     }
 
-    window.closeQuickAddIFrame();
+    window.closeQuickIFrame();
 };
 
-window.closeQuickAddIFrame = function closeQuickAddIFrame(e) {
-    if (e !== undefined) {
-        e.preventDefault();
-    }
+window.closeQuickIFrame = function () {
     $("#create-object-overlay").remove();
 };
+
+function createQuickIframe(url) {
+    window.closeQuickIFrame();
+    const overlay = document.createElement("div");
+    overlay.id = "create-object-overlay";
+
+    const contentPane = document.createElement("div");
+    contentPane.id = "create-object-content-pane";
+    contentPane.className = "content-pane";
+    overlay.appendChild(contentPane);
+
+    const closeIcon = document.createElement("i");
+    closeIcon.className = "fa fa-times-circle-o fa-3x text-danger";
+    const closeButton = document.createElement("a");
+    closeButton.className = "close-btn";
+    closeButton.href = "#";
+    closeButton.onclick = function (e) {
+        if (e !== undefined) {
+            e.preventDefault();
+        }
+        window.closeQuickIFrame();
+    };
+    closeButton.appendChild(closeIcon);
+    contentPane.appendChild(closeButton);
+
+    const iFrame = document.createElement("iframe");
+    iFrame.frameBorder = 0;
+    iFrame.width = "100%";
+    iFrame.height = "100%";
+    iFrame.id = "create-object-iframe";
+
+    iFrame.onload = function() {
+        $("#create-object-content-pane").addClass("open");
+        $("#create-object-iframe").contents().find(".quick-add-btn").remove();
+        $("#create-object-iframe").contents().find(".edit-object-btn").remove();
+    };
+
+    iFrame.setAttribute("src", url);
+    contentPane.appendChild(iFrame);
+    $(document.body).append(overlay);
+}
 
 window.setupQuickAdd = function (element) {
     $(element).on("click", function(e) {
         e.preventDefault();
-        window.closeQuickAddIFrame();
-        const url = $(this).data("url");
-        const overlay = document.createElement("div");
-        overlay.id = "create-object-overlay";
+        createQuickIframe($(this).data("url"));
+    });
+};
 
-        const contentPane = document.createElement("div");
-        contentPane.id = "create-object-content-pane";
-        contentPane.className = "content-pane";
-        overlay.appendChild(contentPane);
+window.setupEditButton = function (element) {
+    // setup each target
+    $(element).each(function () {
+        var target = this;
+        var selectTarget = $(target).data("target");
 
-        const closeIcon = document.createElement("i");
-        closeIcon.className = "fa fa-times-circle-o fa-3x text-danger";
-        const closeButton = document.createElement("a");
-        closeButton.className = "close-btn";
-        closeButton.href = "#";
-        closeButton.onclick = window.closeQuickAddIFrame;
-        closeButton.appendChild(closeIcon);
-        contentPane.appendChild(closeButton);
+        if (selectTarget) {
+            var field = $("[name='" + selectTarget + "']");
 
-        const iFrame = document.createElement("iframe");
-        iFrame.frameBorder = 0;
-        iFrame.width = "100%";
-        iFrame.height = "100%";
-        iFrame.id = "create-object-iframe";
+            if (!field.val()) {
+                $(target).hide();
+            }
 
-        iFrame.onload = function() {
-            $("#create-object-content-pane").addClass("open");
-        };
-
-        iFrame.setAttribute("src", url);
-        contentPane.appendChild(iFrame);
-        $(document.body).append(overlay);
+            field.on("change", function () {
+                if ($(this).val()) {
+                    $(target).show();
+                } else {
+                    $(target).hide();
+                }
+            });
+        }
+    });
+    $(element).on("click", function(e) {
+        e.preventDefault();
+        var selectTarget = $(e.currentTarget).data("target");
+        if (selectTarget) {
+            var field = $("[name='" + selectTarget + "']");
+            var model = $(e.currentTarget).data("edit-model");
+            var selectedValue = $(field).val();
+            if (selectedValue) {
+                var url = window.ShuupAdminConfig.browserUrls.edit + "?mode=iframe&model=" + model + "&id=" + selectedValue;
+                createQuickIframe(url);
+            }
+        }
     });
 };
 
 $(function() {
     window.setupQuickAdd($(".quick-add-btn a.btn"));
+    window.setupEditButton($(".edit-object-btn a.btn"));
 });